requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# OpenTelemetry Core
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0

# OpenTelemetry Instrumentation
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-instrumentation-system-metrics==0.42b0

# OpenTelemetry Exporters
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0

# Mastercard SDK
mastercard-client-encryption==1.2.0
oauth1==1.3.0
cryptography==41.0.7
requests==2.31.0
requests-oauthlib==1.3.1

# Elastic
elasticsearch==8.11.0
elastic-apm==6.19.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
pyyaml==6.0.1

# Logging
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6

